fix: correções na tela de onboarding e reorganização das pastas
