+++ bloom-app/lib/main.dart
@@ -1,7 +1,7 @@
 // main.dart
 
 import 'package:flutter/material.dart';
-import 'package:app/src/modules/auth/login_screen.dart';
+import 'package:app/src/modules/auth/login/login_screen.dart';
 import 'package:app/src/modules/onboarding/onboarding_screen.dart'; // Corrigido o caminho do onboarding
 
 // Variável para forçar o onboarding a ser sempre falso
